streamlit_app/requirements.txt
streamlit>=1.25.0
requests>=2.28.0
aiohttp>=3.8.0
sseclient-py>=1.7.2
python-dotenv>=1.0.0
pillow>=9.0.0 